Refactor build and project services
Move some inappropriate code from the build service
to the project service, along with all dependencies.
